move lexer
