+++ tests/parsers/test_lexer.py
@@ -4,12 +4,12 @@ import os
 import sys
 from collections.abc import Sequence
 
-sys.path.insert(0, os.path.abspath(".."))  # FIXME
+sys.path.insert(0, os.path.abspath("../.."))  # FIXME
 from pprint import pformat
 
 import pytest
 
-from xonsh.lexer import Lexer
+from xonsh.parsers.lexer import Lexer
 from xonsh.ply.ply.lex import LexToken
 
 LEXER_ARGS = {"lextab": "lexer_test_table", "debug": 0}
