+++ xonsh/parsers/base.py
@@ -11,7 +11,7 @@ from collections.abc import Iterable, Mapping, Sequence
 from threading import Thread
 
 from xonsh.lazyasd import LazyObject
-from xonsh.lexer import Lexer, LexToken
+from xonsh.parsers.lexer import Lexer, LexToken
 from xonsh.parsers import ast
 from xonsh.parsers.ast import has_elts, load_attribute_chain, xonsh_call
 from xonsh.parsers.context_check import check_contexts
