+++ xonsh/parsers/completion_context.py
@@ -18,7 +18,7 @@ from typing import (
 )
 
 from xonsh.lazyasd import lazyobject
-from xonsh.lexer import Lexer
+from xonsh.parsers.lexer import Lexer
 from xonsh.parsers.base import Location, raise_parse_error
 from xonsh.ply.ply import yacc
 from xonsh.tools import check_for_partial_string, get_line_continuation
